+++ test/administrative-sdk/basicAuthSpec.js
@@ -16,7 +16,7 @@
 
 require('jasmine-ajax');
 
-const its = require('..');
+const its = require('../../');
 
 describe('BasicAuth object test', function() {
   it('should require all required fields in constructor', function() {
