+++ test/administrative-sdk/organisationSpec.js
@@ -16,33 +16,25 @@
 
 require('jasmine-ajax');
 
-const its = require('..');
+const its = require('../../');
 
-describe('Student object test', function() {
-  it('should instantiate a Student without id', function() {
-    var s = new its.Student();
-    expect(s).toBeDefined();
-    expect(s.id).toBeUndefined();
-    expect(s.organisationId).toBeUndefined();
-    expect(s.firstName).toBeUndefined();
-    expect(s.lastName).toBeUndefined();
-    expect(s.gender).toBeUndefined();
-    expect(s.birthYear).toBeUndefined();
+describe('Organisation object test', function() {
+  it('should instantiate an Organisation without id', function() {
+    var o = new its.Organisation();
+    expect(o).toBeDefined();
+    expect(o.id).toBeUndefined();
+    expect(o.name).toBeUndefined();
   });
 
-  it('should instantiate a Student with id and metadata', function() {
-    var s = new its.Student('fb', 'test', 'Mark', 'Zuckerberg', 'male', 1984);
-    expect(s).toBeDefined();
-    expect(s.id).toBe('test');
-    expect(s.organisationId).toBe('fb');
-    expect(s.firstName).toBe('Mark');
-    expect(s.lastName).toBe('Zuckerberg');
-    expect(s.gender).toBe('male');
-    expect(s.birthYear).toBe(1984);
+  it('should instantiate an Organisation with id and metadata', function() {
+    var o = new its.Organisation('test', 'School of silly walks');
+    expect(o).toBeDefined();
+    expect(o.id).toBe('test');
+    expect(o.name).toBe('School of silly walks');
   });
 });
 
-describe('Student API interaction test', function() {
+describe('Organisation API interaction test', function() {
   beforeEach(function() {
     jasmine.Ajax.install();
   });
@@ -51,31 +43,29 @@ describe('Student API interaction test', function() {
     jasmine.Ajax.uninstall();
   });
 
-  it('should create a new student through API', function() {
-    var student = new its.Student('fb', '1', 'Mark');
+  it('should create a new organisation through API', function() {
+    var organisation = new its.Organisation('1', 'School of silly walks');
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.createStudent(student, cb);
+    var output = api.createOrganisation(organisation, cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations/fb/students';
+    var url = 'https://api.itslanguage.nl/organisations';
     expect(request.url).toBe(url);
     expect(request.method).toBe('POST');
-    var expected = {id: '1',
-      organisationId: 'fb',
-      firstName: 'Mark'};
+    var expected = {id: '1', name: 'School of silly walks'};
     expect(request.data()).toEqual(expected);
 
     var content = {
       id: '1',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      firstName: 'Mark'
+      name: 'School of silly walks'
     };
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 201,
@@ -84,39 +74,37 @@ describe('Student API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    expect(cb).toHaveBeenCalledWith(student);
-    expect(student.id).toBe('1');
-    expect(student.created).toEqual(new Date(stringDate));
-    expect(student.updated).toEqual(new Date(stringDate));
-    expect(student.firstName).toBe('Mark');
+    expect(cb).toHaveBeenCalledWith(organisation);
+    expect(organisation.id).toBe('1');
+    expect(organisation.created).toEqual(new Date(stringDate));
+    expect(organisation.updated).toEqual(new Date(stringDate));
+    expect(organisation.name).toBe('School of silly walks');
   });
 
-  it('should handle errors while creating a new student', function() {
+  it('should handle errors while creating a new organisation', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
-    var student = new its.Student('fb', '1', 'Mark');
+    var organisation = new its.Organisation('1');
     var cb = jasmine.createSpy('callback');
     var ecb = jasmine.createSpy('callback');
 
-    var output = api.createStudent(student, cb, ecb);
+    var output = api.createOrganisation(organisation, cb, ecb);
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations/fb/students';
+    var url = 'https://api.itslanguage.nl/organisations';
     expect(request.url).toBe(url);
     expect(request.method).toBe('POST');
-    var expected = {id: '1',
-      organisationId: 'fb',
-      firstName: 'Mark'};
+    var expected = {id: '1'};
     expect(request.data()).toEqual(expected);
 
     var content = {
       message: 'Validation failed',
       errors: [
         {
-          resource: 'Student',
-          field: 'lastName',
+          resource: 'Organisation',
+          field: 'name',
           code: 'missing'
         }
       ]
@@ -129,26 +117,26 @@ describe('Student API interaction test', function() {
 
     expect(cb).not.toHaveBeenCalled();
     var errors = [{
-      resource: 'Student',
-      field: 'lastName',
+      resource: 'Organisation',
+      field: 'name',
       code: 'missing'
     }];
-    expect(ecb).toHaveBeenCalledWith(errors, student);
+    expect(ecb).toHaveBeenCalledWith(errors, organisation);
     expect(output).toBeUndefined();
   });
 
-  it('should get an existing student through API', function() {
+  it('should get an existing organisation through API', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.getStudent('fb', '4', cb);
+    var output = api.getOrganisation('4', cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations/fb/students/4';
+    var url = 'https://api.itslanguage.nl/organisations/4';
     expect(request.url).toBe(url);
     expect(request.method).toBe('GET');
 
@@ -156,7 +144,7 @@ describe('Student API interaction test', function() {
       id: '4',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      firstName: 'Mark'
+      name: 'School of silly walks'
     };
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 200,
@@ -165,24 +153,24 @@ describe('Student API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    var student = new its.Student('fb', '4', 'Mark');
-    student.created = new Date(stringDate);
-    student.updated = new Date(stringDate);
-    expect(cb).toHaveBeenCalledWith(student);
+    var organisation = new its.Organisation('4', 'School of silly walks');
+    organisation.created = new Date(stringDate);
+    organisation.updated = new Date(stringDate);
+    expect(cb).toHaveBeenCalledWith(organisation);
   });
 
-  it('should get a list of existing students through API', function() {
+  it('should get a list of existing organisations through API', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.listStudents('fb', cb);
+    var output = api.listOrganisations(cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations/fb/students';
+    var url = 'https://api.itslanguage.nl/organisations';
     expect(request.url).toBe(url);
     expect(request.method).toBe('GET');
 
@@ -190,7 +178,7 @@ describe('Student API interaction test', function() {
       id: '4',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      firstName: 'Mark'
+      name: 'School of silly walks'
     }];
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 200,
@@ -199,9 +187,9 @@ describe('Student API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    var student = new its.Student('fb', '4', 'Mark');
-    student.created = new Date(stringDate);
-    student.updated = new Date(stringDate);
-    expect(cb).toHaveBeenCalledWith([student]);
+    var organisation = new its.Organisation('4', 'School of silly walks');
+    organisation.created = new Date(stringDate);
+    organisation.updated = new Date(stringDate);
+    expect(cb).toHaveBeenCalledWith([organisation]);
   });
 });
