+++ test/administrative-sdk/studentSpec.js
@@ -16,25 +16,33 @@
 
 require('jasmine-ajax');
 
-const its = require('..');
+const its = require('../../');
 
-describe('Organisation object test', function() {
-  it('should instantiate an Organisation without id', function() {
-    var o = new its.Organisation();
-    expect(o).toBeDefined();
-    expect(o.id).toBeUndefined();
-    expect(o.name).toBeUndefined();
+describe('Student object test', function() {
+  it('should instantiate a Student without id', function() {
+    var s = new its.Student();
+    expect(s).toBeDefined();
+    expect(s.id).toBeUndefined();
+    expect(s.organisationId).toBeUndefined();
+    expect(s.firstName).toBeUndefined();
+    expect(s.lastName).toBeUndefined();
+    expect(s.gender).toBeUndefined();
+    expect(s.birthYear).toBeUndefined();
   });
 
-  it('should instantiate an Organisation with id and metadata', function() {
-    var o = new its.Organisation('test', 'School of silly walks');
-    expect(o).toBeDefined();
-    expect(o.id).toBe('test');
-    expect(o.name).toBe('School of silly walks');
+  it('should instantiate a Student with id and metadata', function() {
+    var s = new its.Student('fb', 'test', 'Mark', 'Zuckerberg', 'male', 1984);
+    expect(s).toBeDefined();
+    expect(s.id).toBe('test');
+    expect(s.organisationId).toBe('fb');
+    expect(s.firstName).toBe('Mark');
+    expect(s.lastName).toBe('Zuckerberg');
+    expect(s.gender).toBe('male');
+    expect(s.birthYear).toBe(1984);
   });
 });
 
-describe('Organisation API interaction test', function() {
+describe('Student API interaction test', function() {
   beforeEach(function() {
     jasmine.Ajax.install();
   });
@@ -43,29 +51,31 @@ describe('Organisation API interaction test', function() {
     jasmine.Ajax.uninstall();
   });
 
-  it('should create a new organisation through API', function() {
-    var organisation = new its.Organisation('1', 'School of silly walks');
+  it('should create a new student through API', function() {
+    var student = new its.Student('fb', '1', 'Mark');
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.createOrganisation(organisation, cb);
+    var output = api.createStudent(student, cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations';
+    var url = 'https://api.itslanguage.nl/organisations/fb/students';
     expect(request.url).toBe(url);
     expect(request.method).toBe('POST');
-    var expected = {id: '1', name: 'School of silly walks'};
+    var expected = {id: '1',
+      organisationId: 'fb',
+      firstName: 'Mark'};
     expect(request.data()).toEqual(expected);
 
     var content = {
       id: '1',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      name: 'School of silly walks'
+      firstName: 'Mark'
     };
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 201,
@@ -74,37 +84,39 @@ describe('Organisation API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    expect(cb).toHaveBeenCalledWith(organisation);
-    expect(organisation.id).toBe('1');
-    expect(organisation.created).toEqual(new Date(stringDate));
-    expect(organisation.updated).toEqual(new Date(stringDate));
-    expect(organisation.name).toBe('School of silly walks');
+    expect(cb).toHaveBeenCalledWith(student);
+    expect(student.id).toBe('1');
+    expect(student.created).toEqual(new Date(stringDate));
+    expect(student.updated).toEqual(new Date(stringDate));
+    expect(student.firstName).toBe('Mark');
   });
 
-  it('should handle errors while creating a new organisation', function() {
+  it('should handle errors while creating a new student', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
-    var organisation = new its.Organisation('1');
+    var student = new its.Student('fb', '1', 'Mark');
     var cb = jasmine.createSpy('callback');
     var ecb = jasmine.createSpy('callback');
 
-    var output = api.createOrganisation(organisation, cb, ecb);
+    var output = api.createStudent(student, cb, ecb);
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations';
+    var url = 'https://api.itslanguage.nl/organisations/fb/students';
     expect(request.url).toBe(url);
     expect(request.method).toBe('POST');
-    var expected = {id: '1'};
+    var expected = {id: '1',
+      organisationId: 'fb',
+      firstName: 'Mark'};
     expect(request.data()).toEqual(expected);
 
     var content = {
       message: 'Validation failed',
       errors: [
         {
-          resource: 'Organisation',
-          field: 'name',
+          resource: 'Student',
+          field: 'lastName',
           code: 'missing'
         }
       ]
@@ -117,26 +129,26 @@ describe('Organisation API interaction test', function() {
 
     expect(cb).not.toHaveBeenCalled();
     var errors = [{
-      resource: 'Organisation',
-      field: 'name',
+      resource: 'Student',
+      field: 'lastName',
       code: 'missing'
     }];
-    expect(ecb).toHaveBeenCalledWith(errors, organisation);
+    expect(ecb).toHaveBeenCalledWith(errors, student);
     expect(output).toBeUndefined();
   });
 
-  it('should get an existing organisation through API', function() {
+  it('should get an existing student through API', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.getOrganisation('4', cb);
+    var output = api.getStudent('fb', '4', cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations/4';
+    var url = 'https://api.itslanguage.nl/organisations/fb/students/4';
     expect(request.url).toBe(url);
     expect(request.method).toBe('GET');
 
@@ -144,7 +156,7 @@ describe('Organisation API interaction test', function() {
       id: '4',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      name: 'School of silly walks'
+      firstName: 'Mark'
     };
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 200,
@@ -153,24 +165,24 @@ describe('Organisation API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    var organisation = new its.Organisation('4', 'School of silly walks');
-    organisation.created = new Date(stringDate);
-    organisation.updated = new Date(stringDate);
-    expect(cb).toHaveBeenCalledWith(organisation);
+    var student = new its.Student('fb', '4', 'Mark');
+    student.created = new Date(stringDate);
+    student.updated = new Date(stringDate);
+    expect(cb).toHaveBeenCalledWith(student);
   });
 
-  it('should get a list of existing organisations through API', function() {
+  it('should get a list of existing students through API', function() {
     var api = new its.Sdk({
       authPrincipal: 'principal',
       authPassword: 'secret'
     });
     var cb = jasmine.createSpy('callback');
 
-    var output = api.listOrganisations(cb);
+    var output = api.listStudents('fb', cb);
     expect(output).toBeUndefined();
 
     var request = jasmine.Ajax.requests.mostRecent();
-    var url = 'https://api.itslanguage.nl/organisations';
+    var url = 'https://api.itslanguage.nl/organisations/fb/students';
     expect(request.url).toBe(url);
     expect(request.method).toBe('GET');
 
@@ -178,7 +190,7 @@ describe('Organisation API interaction test', function() {
       id: '4',
       created: '2014-12-31T23:59:59Z',
       updated: '2014-12-31T23:59:59Z',
-      name: 'School of silly walks'
+      firstName: 'Mark'
     }];
     jasmine.Ajax.requests.mostRecent().respondWith({
       status: 200,
@@ -187,9 +199,9 @@ describe('Organisation API interaction test', function() {
     });
 
     var stringDate = '2014-12-31T23:59:59Z';
-    var organisation = new its.Organisation('4', 'School of silly walks');
-    organisation.created = new Date(stringDate);
-    organisation.updated = new Date(stringDate);
-    expect(cb).toHaveBeenCalledWith([organisation]);
+    var student = new its.Student('fb', '4', 'Mark');
+    student.created = new Date(stringDate);
+    student.updated = new Date(stringDate);
+    expect(cb).toHaveBeenCalledWith([student]);
   });
 });
