+++ test/audio-toolsSpec.js
@@ -1,0 +1,4 @@
+/* eslint-disable
+ camelcase,
+ new-cap
+ */
