+++ test/web-audio-recorderSpec.js
@@ -1,0 +1,4 @@
+/* eslint-disable
+ camelcase,
+ new-cap
+ */
